Revert "using unsigned int for number of primitives in Embree API"
This reverts commit 5a25a0aaa029bf0f5c771eb25fd6b099a0d69b70.

+++ include/embree2/rtcore_geometry.h
@@ -136,9 +136,9 @@ RTCORE_API void rtcSetTransform (RTCScene scene,                          //!< s
   float used for alignment can be arbitrary. */
 RTCORE_API unsigned rtcNewTriangleMesh (RTCScene scene,                    //!< the scene the mesh belongs to
                                         RTCGeometryFlags flags,            //!< geometry flags
-                                        unsigned int numTriangles,               //!< number of triangles
-                                        unsigned int numVertices,                //!< number of vertices
-                                        unsigned int numTimeSteps = 1            //!< number of motion blur time steps
+                                        size_t numTriangles,               //!< number of triangles
+                                        size_t numVertices,                //!< number of vertices
+                                        size_t numTimeSteps = 1            //!< number of motion blur time steps
   );
 
 /*! \brief Creates a new subdivision mesh. The number of faces
@@ -197,13 +197,13 @@ RTCORE_API unsigned rtcNewTriangleMesh (RTCScene scene,                    //!< 
 */
 RTCORE_API unsigned rtcNewSubdivisionMesh (RTCScene scene,                //!< the scene the mesh belongs to
                                            RTCGeometryFlags flags,        //!< geometry flags
-                                           unsigned int numFaces,               //!< number of faces
-                                           unsigned int numEdges,               //!< number of edges
-                                           unsigned int numVertices,            //!< number of vertices
-                                           unsigned int numEdgeCreases,         //!< number of edge creases
-                                           unsigned int numVertexCreases,       //!< number of vertex creases
-                                           unsigned int numHoles,               //!< number of holes
-                                           unsigned int numTimeSteps = 1        //!< number of motion blur time steps
+                                           size_t numFaces,               //!< number of faces
+                                           size_t numEdges,               //!< number of edges
+                                           size_t numVertices,            //!< number of vertices
+                                           size_t numEdgeCreases,         //!< number of edge creases
+                                           size_t numVertexCreases,       //!< number of vertex creases
+                                           size_t numHoles,               //!< number of holes
+                                           size_t numTimeSteps = 1        //!< number of motion blur time steps
   );
 
 /*! \brief Creates a new hair geometry, consisting of multiple hairs
@@ -226,9 +226,9 @@ RTCORE_API unsigned rtcNewSubdivisionMesh (RTCScene scene,                //!< t
   might show geometric artefacts. */
 RTCORE_API unsigned rtcNewHairGeometry (RTCScene scene,                    //!< the scene the curves belong to
                                         RTCGeometryFlags flags,            //!< geometry flags
-                                        unsigned int numCurves,                  //!< number of curves
-                                        unsigned int numVertices,                //!< number of vertices
-                                        unsigned int numTimeSteps = 1            //!< number of motion blur time steps
+                                        size_t numCurves,                  //!< number of curves
+                                        size_t numVertices,                //!< number of vertices
+                                        size_t numTimeSteps = 1            //!< number of motion blur time steps
   );
 
 /*! \brief Sets 32 bit ray mask. */
